fix comparison between int and LONG_MAX

+++ usr/src/lib/libinetsvc/common/inetsvc.c
@@ -22,6 +22,9 @@
  * Copyright 2009 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
+/*
+ * Copyright 2017 Hayashi Naoyuki
+ */
 
 /*
  * This library contains a set of routines that are shared amongst inetd,
@@ -1305,7 +1308,7 @@ get_rpc_prognum(const char *svc_name)
 {
 	struct rpcent	rpc;
 	char		buf[INETSVC_SVC_BUF_MAX];
-	int		pnum;
+	long		pnum;
 	char		*endptr;
 
 	if (getrpcbyname_r(svc_name, &rpc, buf, sizeof (buf)) != NULL)
@@ -1318,7 +1321,7 @@ get_rpc_prognum(const char *svc_name)
 		return (-1);
 	}
 
-	return (pnum);
+	return (int)(pnum);
 }
 
 /*
